project: add rect progress bar module

+++ garbage_clean/presentation/src/main/java/yin_kio/garbage_clean/presentation/models/UiFileSystemInfo.kt
@@ -4,5 +4,5 @@ data class UiFileSystemInfo(
     val occupied: String = "",
     val available: String = "",
     val total: String = "",
-    val occupiedPercents: Int = 0
+    val occupiedPercents: Float = 0f
 )
